requirements.txt
discord.py
pymongo
pywin32